% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\docType{data}
\name{dp}
\alias{dp}
\title{Data on capital punishment}
\format{A data frame with 17 rows and 7 variables:
\describe{
  \item{EXECUTIONS}{The number of times that capital punishment is implemented on a state level in the United States for the year 1997}
  \item{INCOME}{Median per capita income in dollars}
  \item{PERPOVERTY}{Percent of the population classified as living in poverty}
  \item{PERBLACK}{Percent of Black citizens in the population}
  \item{VC100k96}{Rate of violent crimes per 100,000 residents for the year before (1996)}
  \item{SOUTH}{Dummy variable to indicate whether the state is in the South}
  \item{PROPDEGREE}{Proportion of the population with a college degree}
  ...
}}
\usage{
data(dp)
}
\description{
Data for the capital punishment example used in chapters 4, 5, and 6
}
\examples{
data(dp)
attach(dp)
}
\keyword{datasets}
