% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.summary.multinom.R
\name{glm.summary.multinom}
\alias{glm.summary.multinom}
\title{Summarize regression output from multinomial generalized linear models.}
\usage{
glm.summary.multinom(in.object, alpha = 0.05)
}
\arguments{
\item{in.object}{The regression output from multinom().}

\item{alpha}{A parameter defaulted to 0.05.}
}
\value{
The output is a list.
}
\description{
An alternative to the summary() function.
}
\examples{
data(primary)
attach(primary)
library(nnet)
primary.out <- multinom(PRIMVOTE ~ AGE + GENDER + EDUCATION + REGION +
                         RELIGIOSITY + IDEOLOGY + RWA + TRUMPWIN, data=primary)
glm.summary.multinom(primary.out)

}
